+++ packages/core/src/__tests__/builder.test.ts
@@ -13,6 +13,8 @@ import customCode from './data/blocks/custom-code.raw.tsx?raw';
 import embed from './data/blocks/embed.raw.tsx?raw';
 import image from './data/blocks/image.raw.tsx?raw';
 import stamped from './data/blocks/stamped-io.raw.tsx?raw';
+import booleanContent from './data/builder/boolean.json?raw';
+import customComponentSlotPropertyContent from './data/builder/custom-component-slot-property.json?raw';
 import lazyLoadSection from './data/builder/lazy-load-section.json?raw';
 import slotsContent from './data/builder/slots.json?raw';
 import slots2Content from './data/builder/slots2.json?raw';
@@ -366,6 +368,30 @@ describe('Builder', () => {
 
   test('slots2', async () => {
     const component = builderContentToMitosisComponent(JSON.parse(slots2Content));
+
+    const out = await componentToReact({
+      plugins: [compileAwayBuilderComponents()],
+    })({ component });
+
+    expect(component).toMatchSnapshot();
+    expect(out).toMatchSnapshot();
+  });
+
+  test('customComponentSlotProperty', async () => {
+    const component = builderContentToMitosisComponent(
+      JSON.parse(customComponentSlotPropertyContent),
+    );
+
+    const out = await componentToReact({
+      plugins: [compileAwayBuilderComponents()],
+    })({ component });
+
+    expect(component).toMatchSnapshot();
+    expect(out).toMatchSnapshot();
+  });
+
+  test('booleans', async () => {
+    const component = builderContentToMitosisComponent(JSON.parse(booleanContent));
 
     const out = await componentToReact({
       plugins: [compileAwayBuilderComponents()],